MemoryRegion: Use ints instead of ushorts for addresses

+++ GbSharp/Memory/GbMemory.cs
@@ -21,13 +21,13 @@ namespace GbSharp.Memory
         // 0xFF80 to 0xFFFE: High Speed RAM
         // 0xFFFF          : IE Register
 
-        private readonly Dictionary<ushort, MemoryRegion> MemoryMap;
+        private readonly Dictionary<int, MemoryRegion> MemoryMap;
         private BootRomRegion BootRomRegion;
         private bool BootRomAccessible;
 
         public GbMemory()
         {
-            MemoryMap = new Dictionary<ushort, MemoryRegion>();
+            MemoryMap = new Dictionary<int, MemoryRegion>();
             BootRomAccessible = false;
 
             WorkRamRegion workRam = new WorkRamRegion();
@@ -40,15 +40,15 @@ namespace GbSharp.Memory
             RegisterMmio(0xFF50, () => (byte)(BootRomAccessible ? 0 : 1), x => BootRomAccessible = false);
         }
 
-        public void RegisterRegion(ushort address, int size, MemoryRegion region)
+        public void RegisterRegion(int address, int size, MemoryRegion region)
         {
             for (int i = 0; i < size; i++)
             {
-                MemoryMap[(ushort)(address + i)] = region;
+                MemoryMap[address + i] = region;
             }
         }
 
-        public void RegisterMmio(ushort address, Func<byte> readFunc, Action<byte> writeFunc)
+        public void RegisterMmio(int address, Func<byte> readFunc, Action<byte> writeFunc)
         {
             MemoryMap[address] = new MmioRegion(readFunc, writeFunc);
         }
@@ -59,7 +59,7 @@ namespace GbSharp.Memory
             BootRomAccessible = true;
         }
 
-        public byte Read(ushort address)
+        public byte Read(int address)
         {
             if (BootRomAccessible)
             {
@@ -80,7 +80,7 @@ namespace GbSharp.Memory
             }
         }
 
-        public void Write(ushort address, byte val)
+        public void Write(int address, byte val)
         {
             if (MemoryMap.TryGetValue(address, out MemoryRegion region))
             {
