+++ GbSharp/Ppu/Memory/VideoRamRegion.cs
@@ -31,7 +31,7 @@ namespace GbSharp.Ppu.Memory
             Locked = false;
         }
 
-        public override byte Read(ushort address)
+        public override byte Read(int address)
         {
             // Reading VRAM from the CPU during certain modes is disallowed
             if (Locked)
@@ -39,24 +39,24 @@ namespace GbSharp.Ppu.Memory
                 return 0xFF;
             }
 
-            return ReadDirect((ushort)(address & 0x1FFF));
+            return ReadDirect(address & 0x1FFF);
         }
 
-        public override void Write(ushort address, byte val)
+        public override void Write(int address, byte val)
         {
             // Writes to VRAM from the CPU during certain modes are ignored
             if (!Locked)
             {
-                WriteDirect((ushort)(address & 0x1FFF), val);
+                WriteDirect(address & 0x1FFF, val);
             }
         }
 
-        public byte ReadDirect(ushort offset)
+        public byte ReadDirect(int offset)
         {
             return Banks[CurrentSwitchableBank][offset];
         }
 
-        public void WriteDirect(ushort offset, byte val)
+        public void WriteDirect(int offset, byte val)
         {
             Banks[CurrentSwitchableBank][offset] = val;
         }
